Pending specs for data searches.
The version of ApacheDS that's currently embedded is not compatible with
ruby-net-ldap.  Problem seems to be a combination of Net::LDAP including more
in its search request than is necessary and a glaring off-by-one bug in
ApacheDS's org.apache.directory.shared.ldap.codec.search.controls.pagedSearch.PagedSearchControlStatesEnum.
Will need to resolve this before Ladle can be useful.