Ignore no-such-process when stopping.
Later versions of JRuby throw ESRCH when you kill a process that is
already gone. CRuby's Process.kill does this when you just invent a
pid, but does not do it in this specific context. I theorize that
this is because the pid is a former child of the killing process, but
I don't know.