Correct `Before` example for cucumber. Closes #8.
The previous example did keep the server running for all tests, but
since it kept track of whether it had been started in an instance
variable, the second and every subsequent test would try to start a
new server. This would fail since the original one was still running
on the configured port.

Thanks to @alanfranz for identifying this problem.